Add new logic to allow soft handling of encoding read errors with user ability to hard-fail
result of *.reg inclusion that caused a failure requiring that type to be excluded.  To error on side of best practice, don't outright fail as we are only told what directory to read and encoding to read it in as.  If the encoding is wrong, we cannot really address that here so log the files and continue.  If user chooses, they can cause build to fail.